+++ src/Documentation.ts
@@ -1,21 +1,21 @@
 import { type Editor } from "./main";
-import { introduction } from "./documentation/introduction";
+import { introduction } from "./documentation/basics/welcome";
+import { software_interface } from "./documentation/basics/interface";
+import { shortcuts } from "./documentation/basics/keyboard";
+import { code } from "./documentation/basics/code";
+import { mouse } from "./documentation/basics/mouse";
 import { oscilloscope } from "./documentation/oscilloscope";
 import { synchronisation } from "./documentation/synchronisation";
 import { samples } from "./documentation/samples";
 import { chaining } from "./documentation/chaining";
-import { software_interface } from "./documentation/interface";
 import { interaction } from "./documentation/interaction";
 import { time } from "./documentation/time";
 import { linear_time } from "./documentation/linear_time";
 import { cyclical_time } from "./documentation/cyclical_time";
 import { long_forms } from "./documentation/long_forms";
 import { midi } from "./documentation/midi";
-import { code } from "./documentation/code";
 import { about } from "./documentation/about";
 import { sound } from "./documentation/engine";
-import { shortcuts } from "./documentation/keyboard";
-import { mouse } from "./documentation/mouse";
 import { patterns } from "./documentation/patterns";
 import { functions } from "./documentation/functions";
 import { variables } from "./documentation/variables";
