+++ src/documentation/basics/welcome.ts
@@ -1,6 +1,6 @@
-import { makeExampleFactory, key_shortcut } from "../Documentation";
-import { type Editor } from "../main";
-import { examples } from "../examples/excerpts";
+import { makeExampleFactory, key_shortcut } from "../../Documentation";
+import { type Editor } from "../../main";
+import { examples } from "../../examples/excerpts";
 
 export const introduction = (application: Editor): string => {
   const makeExample = makeExampleFactory(application);
